Core Wifi Support complete(disabled) See #define ENABLE_WIFI_SUPPORT

+++ SIO_ESP32WRM_Relay_X2/webPages.h
@@ -42,30 +42,30 @@ void storeWifiScanResult(int networksFound) {
       Networks += dBmtoPercentage(WiFi.RSSI(i));
       Networks += "%)\"";
     }
-    #ifdef _GDebug
-    Serial.printf("%d: %s, Ch:%d (%ddBm) %s\n", i + 1, WiFi.SSID(i).c_str(), WiFi.channel(i), WiFi.RSSI(i), WiFi.encryptionType(i) == WIFI_AUTH_OPEN ? "open" : "");
-    #endif
+    if(_debug){
+      debugMsgPrefx();Serial.printf("%d: %s, Ch:%d (%ddBm) %s\n", i + 1, WiFi.SSID(i).c_str(), WiFi.channel(i), WiFi.RSSI(i), WiFi.encryptionType(i) == WIFI_AUTH_OPEN ? "open" : "");
+    }
     
   }
   
   Networks += "]}";
   
-  //WiFi.scanDelete(); not supported for esp8266
+  WiFi.scanDelete(); //not supported for esp8266
   
 }
 
-const char* www_realm = "Login Required";
+const char* www_realm = "admin@SIOControl";
 String authFailResponse = "Authentication Failed";
 bool checkAuth(){
   
   if(!webServer.authenticate(www_username,wifiConfig.apPassword)){
     #ifdef USE_DIGESTAUTH
     webServer.requestAuthentication(DIGEST_AUTH, www_realm, authFailResponse);
-      #ifdef _GDebug
-      debugMsgPrefx();
-      Serial.print("AuthFailed?:");
-      Serial.println(authFailResponse);
-      #endif
+      if(_debug){
+        debugMsgPrefx();
+        Serial.print("AuthFailed?:");
+        Serial.println(authFailResponse);
+      }
     #else
     webServer.requestAuthentication();
     #endif
@@ -81,16 +81,16 @@ bool handleFileRead(String path) {
   if(!checkAuth())
     return false;
     
-  #ifdef _GDebug
-  debugMsgPrefx();Serial.println("start handleFileRead");
-  #endif
+  if(_debug){
+    debugMsgPrefx();Serial.println("start handleFileRead");
+  }
   
   if (SPIFFS.exists(path)) {
-    #ifdef _GDebug
-    debugMsgPrefx();
-    Serial.print("File found:");
-    Serial.println(path);
-    #endif
+    if(_debug){
+      debugMsgPrefx();
+      Serial.print("File found:");
+      Serial.println(path);
+    }
     File file = SPIFFS.open(path, "r");
     if(path.endsWith("css")){
       webServer.streamFile(file, "text/css");
@@ -100,19 +100,19 @@ bool handleFileRead(String path) {
     file.close();
     return true;
   }else{
-    #ifdef _GDebug
-    debugMsgPrefx();
-    Serial.print("File not found");
-    Serial.println(path);
-    #endif
+    if(_debug){
+      debugMsgPrefx();
+      Serial.print("File not found");
+      Serial.println(path);
+    }
   }
   return false;
 }
 
 void initialisePages(){
-  #ifdef _GDebug
-  debugMsgPrefx();Serial.println("Initializing pages start");
-  #endif
+  if(_debug){
+    debugMsgPrefx();Serial.println("Initializing pages start");
+  }
   webServer.serveStatic("/config.js",SPIFFS,"/config.js");//.setAuthentication(www_username,wifiConfig.apPassword);
   webServer.serveStatic("/ioconfig.js",SPIFFS,"/ioconfig.js");//.setAuthentication(www_username,wifiConfig.apPassword);
   webServer.serveStatic("/style.css",SPIFFS,"/style.css");//.setAuthentication(www_username,wifiConfig.apPassword);//31536000
@@ -120,27 +120,27 @@ void initialisePages(){
   
   
   webServer.on("/config",HTTP_GET,[]{
-    #ifdef _GDebug
-    debugMsg("start config");
-    #endif
+    if(_debug){
+      debugMsg("start config");
+    }
     
     handleFileRead("/config.htm");
     
-    #ifdef _GDebug
-    debugMsg("end config");
-    #endif
+    if(_debug){
+      debugMsg("end config");
+    }
   });
 
   webServer.on("/ioconfig",HTTP_GET,[]{
-    #ifdef _GDebug
-    debugMsg("start ioconfig");
-    #endif
+    if(_debug){
+      debugMsg("start ioconfig");
+    }
     
     handleFileRead("/ioconfig.htm");
     
-    #ifdef _GDebug
-    debugMsg("end ioconfig");
-    #endif
+    if(_debug){
+      debugMsg("end ioconfig");
+    }
   });
   
   webServer.on("/APIGetIOSettings",HTTP_GET,[]{
@@ -207,23 +207,23 @@ void initialisePages(){
     int n = WiFi.scanNetworks();
     storeWifiScanResult(n);
     doWiFiScan = true;
-    #ifdef _GDebug
-    debugMsg("Network Scan requested");
-    #endif
+    if(_debug){
+      debugMsg("Network Scan requested");
+    }
     webServer.send(200, "text/plain","OK" );
   });
 
   webServer.on("/APIGetNetworks",HTTP_GET,[]{
   if(!checkAuth()){return;}
     if(Networks.length() == 0){
-      #ifdef _GDebug
-      debugMsg("Sending Empty Network list");
-      #endif
+      if(_debug){
+        debugMsg("Sending Empty Network list");
+      }
       webServer.send(200, "application/json", "{\"networks\":[]}");
     }else{
-      #ifdef _GDebug
-      debugMsg("Sending Network list");
-      #endif
+      if(_debug){
+        debugMsg("Sending Network list");
+      }
       
       webServer.send(200, "application/json", Networks);
       Networks = "";  
@@ -234,9 +234,9 @@ void initialisePages(){
     if(!checkAuth()){return;}
 
     if(webServer.hasArg("update")){
-      #ifdef _GDebug
-      debugMsg("updating settings");
-      #endif
+      if(_debug){
+        debugMsg("updating settings");
+      }
 
       strlcpy(wifiConfig.wifimode, webServer.arg("cbo_WFMD").c_str(), sizeof(wifiConfig.wifimode));
       strlcpy(wifiConfig.hostname, webServer.arg("tx_HNAM").c_str(), sizeof(wifiConfig.hostname));
@@ -247,13 +247,14 @@ void initialisePages(){
       strlcpy(wifiConfig.apPassword, webServer.arg("tx_APW").c_str(), sizeof(wifiConfig.apPassword));
       
       
-      #ifdef _GDebug
-      debugMsg("AfterUpdate: wificonfig");
-      DebugwifiConfig();
-      #endif
+      if(_debug){
+        debugMsg("AfterUpdate: wificonfig");
+        DebugwifiConfig();
+        debugMsg("before direct Update: SettingsConfig");
+        DebugSettingsConfig(); //prints the settings struct to the Serial line
+      }
       
-      debugMsg("before direct Update: SettingsConfig");
-      DebugSettingsConfig(); //prints the settings struct to the Serial line
+      
       
       //MQTT settings (future)
       /*
@@ -283,9 +284,9 @@ void initialisePages(){
       LastStatus = "Update Completed(Ready!)";
       
     }else if(webServer.hasArg("reboot")){
-      #ifdef _GDebug
-      debugMsg("web called a reboot");
-      #endif
+      if(_debug){
+        debugMsg("web called a reboot");
+      }
       
       LastStatus = "Rebooting in 5 Sec";
       resetTimeDelay.set(10000UL); //need to give it more time so that the web page shows the message.
@@ -294,9 +295,9 @@ void initialisePages(){
     }
     
     handleFileRead("/config.htm");
-    #ifdef _GDebug
-    debugMsg("end post config");
-    #endif
+    if(_debug){
+      debugMsg("end post config");
+    }
     
   });
   
@@ -304,19 +305,19 @@ void initialisePages(){
     if(!checkAuth()){return;}
 
     if(webServer.hasArg("update")){
-      #ifdef _GDebug
-      debugMsg("updating ioconfig");
-      #endif  
+      if(_debug){
+        debugMsg("updating ioconfig");
+      } 
       
       String newIOConfig = "";
       for (int i=0;i<IOSize;i++){      
         newIOConfig += webServer.arg("cbo_io"+String(i)+"_type");
       }
 
-      #ifdef _GDebug
-      debugMsgPrefx();Serial.print("New IO String:");
-      Serial.println(newIOConfig);
-      #endif
+      if(_debug){
+        debugMsgPrefx();Serial.print("New IO String:");
+        Serial.println(newIOConfig);
+      }
 
       updateIOConfig(newIOConfig);
       StoreIOConfig();
@@ -325,15 +326,17 @@ void initialisePages(){
     
     handleFileRead("/ioconfig.htm");
   
-    #ifdef _GDebug
-    debugMsg("end post config");
-    #endif
+    if(_debug){
+      debugMsg("end post config");
+    }
     
   });
   
   webServer.on("/ioControl/status",HTTP_GET,[]{
     if(!checkAuth()){return;}
-
+    if(_debug){
+      debugMsg("[net] ioControl/status Called");
+    }
     String ioData = "{\"IO\":{";
     ioData += "\"io0\":\"" + String(IO[0]) + "\""; //initial item (no comma)      
     for (int i=1;i<IOSize;i++){
